+++ Source/MethodCallReturn.cs
@@ -100,14 +100,7 @@ namespace Moq
 				return this.Returns(() => (TResult)(object)valueFunction);
 			}
 
-			// The following may seem overly cautious, but we don't throw an `ArgumentNullException`
-			// here because Moq has been very forgiving with incorrect `Returns` in the past.
-			if (valueFunction == null)
-			{
-				return this.Returns(() => default(TResult));
-			}
-
-			if (valueFunction.GetMethodInfo().ReturnType == typeof(void))
+			if (valueFunction != null && valueFunction.GetMethodInfo().ReturnType == typeof(void))
 			{
 				throw new ArgumentException(Resources.InvalidReturnsCallbackNotADelegateWithReturnType, nameof(valueFunction));
 			}
@@ -154,12 +147,22 @@ namespace Moq
 
 		private void SetReturnDelegate(Delegate value)
 		{
-			if (value != null)
+			if (value == null)
+			{
+				// A `null` reference (instead of a valid delegate) is interpreted as the actual return value.
+				// This is necessary because the compiler might have picked the unexpected overload for calls like `Returns(null)`,
+				// or the user might have picked an overload like `Returns<T>(null)`,
+				// and instead of in `Returns(TResult)`, we ended up in `Returns(Delegate)` or `Returns(Func)`,
+				// which likely isn't what the user intended.
+				// So here we do what we would've done in `Returns(TResult)`:
+				this.valueDel = (Func<TResult>)(() => default(TResult));
+			}
+			else
 			{
 				ValidateReturnDelegate(value);
+				this.valueDel = value;
 			}
 
-			this.valueDel = value;
 			this.returnValueKind = ReturnValueKind.Explicit;
 		}